NEX-213: - integration FullBindableOrder in DTO

+++ Next2/src/Next2/Converters/FloatToBoolConverter.cs
@@ -8,7 +8,7 @@ namespace Next2.Converters
     {
         public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
         {
-            return (float)value != 0;
+            return (decimal)value != 0;
         }
 
         public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
