+++ Next2/src/Next2/ViewModels/Mobile/InputGiftCardPageViewModel.cs
@@ -46,7 +46,7 @@ namespace Next2.ViewModels.Mobile
 
         public string InputGiftCardFounds { get; set; }
 
-        public float RemainingGiftCardTotal { get; set; }
+        public decimal RemainingGiftCardTotal { get; set; }
 
         public bool IsInSufficientGiftCardFunds { get; set; }
 
@@ -80,7 +80,7 @@ namespace Next2.ViewModels.Mobile
                 {
                     RemainingGiftCardTotal = Customer.GiftCardTotal;
 
-                    if (float.TryParse(InputGiftCardFounds, out float sum))
+                    if (decimal.TryParse(InputGiftCardFounds, out decimal sum))
                     {
                         sum /= 100;
 
@@ -97,7 +97,7 @@ namespace Next2.ViewModels.Mobile
                 }
                 else
                 {
-                    if (float.TryParse(InputGiftCardFounds, out float sum))
+                    if (decimal.TryParse(InputGiftCardFounds, out decimal sum))
                     {
                         IsInSufficientGiftCardFunds = RemainingGiftCardTotal < sum;
                     }
@@ -128,7 +128,7 @@ namespace Next2.ViewModels.Mobile
 
                 if (parameters.ContainsKey(Constants.DialogParameterKeys.GIFT_CARD_ADDED))
                 {
-                    if (float.TryParse(InputGiftCardFounds, out float sum))
+                    if (decimal.TryParse(InputGiftCardFounds, out decimal sum))
                     {
                         sum /= 100;
