+++ Next2/src/Next2/ViewModels/PaymentCompleteViewModel.cs
@@ -31,7 +31,7 @@ namespace Next2.ViewModels
 
         private ICommand _tapTipItemCommand;
 
-        private float _subtotalWithBonus;
+        private decimal _subtotalWithBonus;
 
         public PaymentCompleteViewModel(
             INavigationService navigationService,
@@ -126,7 +126,7 @@ namespace Next2.ViewModels
 
             if(args.PropertyName == nameof(InputValue))
             {
-                if (float.TryParse(InputValue, out float sum))
+                if (decimal.TryParse(InputValue, out decimal sum))
                 {
                     Order.Cash = sum / 100;
                 }
@@ -146,7 +146,7 @@ namespace Next2.ViewModels
                     Order.GiftCard = 0;
                     Order.RemainingGiftCardsTotalFunds = Order.Customer.GiftCardTotal;
 
-                    if (float.TryParse(InputGiftCardFounds, out float sum))
+                    if (decimal.TryParse(InputGiftCardFounds, out decimal sum))
                     {
                         sum /= 100;
 
@@ -173,7 +173,7 @@ namespace Next2.ViewModels
                 }
                 else
                 {
-                    if (float.TryParse(InputGiftCardFounds, out float sum))
+                    if (decimal.TryParse(InputGiftCardFounds, out decimal sum))
                     {
                         IsInsufficientGiftCardFunds = sum > 0;
                     }
@@ -182,7 +182,7 @@ namespace Next2.ViewModels
 
             if (args.PropertyName == nameof(InputTip))
             {
-                if (float.TryParse(InputTip, out float tip))
+                if (decimal.TryParse(InputTip, out decimal tip))
                 {
                     Order.Tip = tip / 100;
                     SelectedTipItem.Text = LocalizationResourceManager.Current["CurrencySign"] + $" {Order.Tip:F2}";
@@ -257,28 +257,28 @@ namespace Next2.ViewModels
                 {
                     TipType = ETipType.NoTip,
                     Text = LocalizationResourceManager.Current["NoTip"],
-                    PercentTip = 0f,
+                    PercentTip = 0m,
                 },
                 new()
                 {
                     TipType = ETipType.Percent,
-                    PercentTip = 0.1f,
+                    PercentTip = 0.1m,
                 },
                 new()
                 {
                     TipType = ETipType.Percent,
-                    PercentTip = 0.15f,
+                    PercentTip = 0.15m,
                 },
                 new()
                 {
                     TipType = ETipType.Percent,
-                    PercentTip = 0.2f,
+                    PercentTip = 0.2m,
                 },
                 new()
                 {
                     TipType = ETipType.Other,
                     Text = LocalizationResourceManager.Current["Other"],
-                    Value = 0f,
+                    Value = 0m,
                 },
             };
 
@@ -421,7 +421,7 @@ namespace Next2.ViewModels
                         Order.GiftCardsTotalFunds = Order.Customer.GiftCardTotal;
                         Order.RemainingGiftCardsTotalFunds = Order.GiftCardsTotalFunds;
 
-                        if (float.TryParse(InputGiftCardFounds, out float sum))
+                        if (decimal.TryParse(InputGiftCardFounds, out decimal sum))
                         {
                             sum /= 100;
 
@@ -452,7 +452,7 @@ namespace Next2.ViewModels
 
         private void RecalculateCustomerGiftCardFounds(CustomerModel customer)
         {
-            float totalPrice = Order.GiftCard;
+            decimal totalPrice = Order.GiftCard;
 
             foreach (var giftCard in customer.GiftCards)
             {
